requirements.txt
requests
beautifulsoup4
tqdm
pandas
psycopg2-binary
boto3
python-dotenv